requirements.txt
flask
pyjwt
werkzeug
Flask-SQLAlchemy
requests
beautifulsoup4
gunicorn
